Refactor useNamespaceFactory imports from CommunityStakeStep to CommunityTransactions across multiple components
