+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/index.ts
@@ -2,5 +2,6 @@ export { default } from './CommunityTransactions';
 export { default as SignCommunityTransactions } from './SignCommunityTransactions';
 export * from './transactionUtils';
 export * from './types';
+export { default as useNamespaceFactory } from './useNamespaceFactory';
 export { default as useNamespaceTransaction } from './useNamespaceTransaction';
 export { default as useStakeTransaction } from './useStakeTransaction';
